[refactor] allow the user to select whether to create or load a project on init

+++ src/odemis/gui/cont/tabs/cryo_chamber_tab.py
@@ -90,8 +90,9 @@ class CryoChamberTab(Tab):
         self.txt_projectpath = self.panel.txt_projectpath
         self.btn_load_project.Hide()
 
-        # Create new project directory on starting the GUI
-        self._create_new_dir()
+        # indicates whether a project has been initially created or loaded
+        self._is_initial_project_ready: bool = False
+
         self._cancel = False
 
         self._current_position = UNKNOWN  # position of the sample (regularly updated)
@@ -260,15 +261,16 @@ class CryoChamberTab(Tab):
         Shows a dialog to change the path and name of the project directory.
         returns nothing, but updates .conf and project path text control
         """
-        # TODO: do not warn if there is no data (eg, at init)
-        box = wx.MessageDialog(self.main_frame,
-                               "This will clear the current project data from Odemis",
-                               caption="Reset Project", style=wx.YES_NO | wx.ICON_QUESTION | wx.CENTER)
+        # do not warn if there is no data (eg, at init)
+        if self._is_initial_project_ready:
+            box = wx.MessageDialog(self.main_frame,
+                                "This will clear the current project data from Odemis",
+                                caption="Reset Project", style=wx.YES_NO | wx.ICON_QUESTION | wx.CENTER)
 
-        box.SetYesNoLabels("&Reset Project", "&Cancel")
-        ans = box.ShowModal()  # Waits for the window to be closed
-        if ans == wx.ID_NO:
-            return
+            box.SetYesNoLabels("&Reset Project", "&Cancel")
+            ans = box.ShowModal()  # Waits for the window to be closed
+            if ans == wx.ID_NO:
+                return
 
         prev_dir = self.conf.pj_last_path
 
@@ -297,29 +299,33 @@ class CryoChamberTab(Tab):
             dlg.Destroy()
             return
 
-        # Reset project, clear the data
-        self._reset_project_data()
+        # only reset and delete data if not initial setup
+        if self._is_initial_project_ready:
+            # Reset project, clear the data
+            self._reset_project_data()
 
-        # If the previous project is empty, it means the user never used it.
-        # (for instance, this happens just after starting the GUI and the user
-        # doesn't like the automatically chosen name)
-        # => Just automatically delete previous folder if empty.
-        try:
-            if os.path.isdir(prev_dir) and not os.listdir(prev_dir):
-                logging.debug("Deleting empty project folder %s", prev_dir)
-                os.rmdir(prev_dir)
-        except Exception:
-            # It might be just due to some access rights, let's not worry too much
-            logging.exception("Failed to delete previous project folder %s", prev_dir)
+            # If the previous project is empty, it means the user never used it.
+            # (for instance, this happens just after starting the GUI and the user
+            # doesn't like the automatically chosen name)
+            # => Just automatically delete previous folder if empty.
+            try:
+                if os.path.isdir(prev_dir) and not os.listdir(prev_dir):
+                    logging.debug("Deleting empty project folder %s", prev_dir)
+                    os.rmdir(prev_dir)
+            except Exception:
+                # It might be just due to some access rights, let's not worry too much
+                logging.exception("Failed to delete previous project folder %s", prev_dir)
 
-        # Handle weird cases where the previous directory would point to the same
-        # folder as the new folder, either with completely the same path, or with
-        # different paths (eg, due to symbolic links).
-        if not os.path.isdir(new_dir):
-            logging.warning("Recreating folder %s which was gone", new_dir)
-            os.mkdir(new_dir)
+            # Handle weird cases where the previous directory would point to the same
+            # folder as the new folder, either with completely the same path, or with
+            # different paths (eg, due to symbolic links).
+            if not os.path.isdir(new_dir):
+                logging.warning("Recreating folder %s which was gone", new_dir)
+                os.mkdir(new_dir)
 
         self._change_project_conf(new_dir)
+
+        return True
 
     def _change_project_conf(self, new_dir):
         """
@@ -353,6 +359,12 @@ class CryoChamberTab(Tab):
 
         self._change_project_conf(np)
 
+    def _create_project(self):
+        """
+        Create a project based on the user selected directory and name
+        """
+        return self._on_change_project_folder(None)
+
     def _reset_project_data(self):
         try:
             streams = self._get_overview_view().getStreams()
@@ -370,15 +382,16 @@ class CryoChamberTab(Tab):
 
     def _load_project_data(self, evt: wx.Event):
 
-        # TODO: do not warn if there is no data (eg, at init)
-        box = wx.MessageDialog(self.main_frame,
-                               "This will clear the current project data from Odemis",
-                               caption="Reset Project", style=wx.YES_NO | wx.ICON_QUESTION | wx.CENTER)
+        # do not warn if there is no data (eg, at init)
+        if self._is_initial_project_ready:
+            box = wx.MessageDialog(self.main_frame,
+                                "This will clear the current project data from Odemis",
+                                caption="Reset Project", style=wx.YES_NO | wx.ICON_QUESTION | wx.CENTER)
 
-        box.SetYesNoLabels("&Reset Project", "&Cancel")
-        ans = box.ShowModal()  # Waits for the window to be closed
-        if ans == wx.ID_NO:
-            return
+            box.SetYesNoLabels("&Reset Project", "&Cancel")
+            ans = box.ShowModal()  # Waits for the window to be closed
+            if ans == wx.ID_NO:
+                return
 
         # select a project directory to load
         proj_path = LoadProjectFileDialog(self.panel, self.conf.pj_last_path)
@@ -388,7 +401,8 @@ class CryoChamberTab(Tab):
         logging.debug("Selected project folder %s", proj_path)
 
         # Reset project, clear the data
-        self._reset_project_data()
+        if self._is_initial_project_ready:
+            self._reset_project_data()
 
         # dont delete empty directories for now
         # If the previous project is empty, it means the user never used it.
@@ -417,7 +431,7 @@ class CryoChamberTab(Tab):
 
             box.SetOKLabel("&OK")
             ans = box.ShowModal()  # Waits for the window to be closed
-            return
+            return True
 
         # follow the order of the data loading in the localization tab
         # load overview streams
@@ -441,7 +455,7 @@ class CryoChamberTab(Tab):
         logging.debug(f"{len(proj_data['overviews'])} overviews loaded.")
         logging.debug(f"{len(localization_tab.tab_data_model.streams.value)} streams loaded.")
 
-        return
+        return True
 
     @call_in_wx_main
     def _update_progress_bar(self, pos):
@@ -909,6 +923,37 @@ class CryoChamberTab(Tab):
 
     def Show(self, show=True):
         Tab.Show(self, show=show)
+
+        # the first time the tab is shown, ask the user to create or load a project
+        if self.IsShown() and not self._is_initial_project_ready:
+
+            # stay in the loop until the user successfully creates or loads a project
+            while True:
+
+                # Ask the user to create or load a project
+                box = wx.MessageDialog(
+                    self.main_frame,
+                    message="Create or Load a Meteor Project?",
+                    caption="Meteor",
+                    style=wx.YES_NO | wx.ICON_QUESTION | wx.CENTER,
+                )
+
+                box.SetYesNoLabels("&Load Project", "&Create Project",)
+                ans = box.ShowModal()  # Waits for the window to be closed
+
+                if ans == wx.ID_NO:
+                    logging.debug("creating project")
+                    ret = self._create_project()
+                if ans == wx.ID_YES:
+                    logging.debug("loading project")
+                    ret = self._load_project_data(None)
+
+                # break if the user successfully created or loaded a project
+                if ret:
+                    break
+
+            # Set the project as ready
+            self._is_initial_project_ready = True
 
     def query_terminate(self):
         """
